pyproject.toml
[project]
name = "BuildingWithLLM"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.52.0",
    "python-dotenv>=1.0.0",
    "asyncpg==0.31.0",
    "fastapi[standard]==0.115.12",
    "sqlalchemy==2.0.39",
    "alembic==1.17.2",
    "greenlet==3.3.0",
    "pgvector>=0.3.0",
    "matplotlib>=3.8.0",
]